% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataHandling.R
\name{DataHandling}
\alias{DataHandling}
\title{DataHandling R6 Class}
\value{
An R6 object of class DataHandling
}
\description{
Class for handling data import and transformation from Excel files.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataHandling-new}{\code{DataHandling$new()}}
\item \href{#method-DataHandling-transform_xlsx}{\code{DataHandling$transform_xlsx()}}
\item \href{#method-DataHandling-server}{\code{DataHandling$server()}}
\item \href{#method-DataHandling-clone}{\code{DataHandling$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataHandling-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataHandling-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHandling$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new instance of DataHandling
Transform Excel file into SummarizedExperiment

Reads an Excel file with sheets "Samples", "Construct_Metadata", and "Construct_Counts" and
converts it into a SummarizedExperiment object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataHandling-transform_xlsx"></a>}}
\if{latex}{\out{\hypertarget{method-DataHandling-transform_xlsx}{}}}
\subsection{Method \code{transform_xlsx()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHandling$transform_xlsx(table_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table_path}}{Character string. Path to the Excel file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A SummarizedExperiment object containing assays (counts), colData (samples), and rowData (construct metadata).
Server logic placeholder
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataHandling-server"></a>}}
\if{latex}{\out{\hypertarget{method-DataHandling-server}{}}}
\subsection{Method \code{server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHandling$server(input, output)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{Shiny input object.}

\item{\code{output}}{Shiny output object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataHandling-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataHandling-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataHandling$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
