% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialExpression.R
\name{DifferentialExpression}
\alias{DifferentialExpression}
\title{DifferentialExpression Class}
\value{
An R6 object of class DifferentialExpression.
}
\description{
This class handles differential expression analysis.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dataHandling}}{An object of class DataHandling used to manage data input.}

\item{\code{test_mode_3}}{Logical flag for test mode.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DifferentialExpression-new}{\code{DifferentialExpression$new()}}
\item \href{#method-DifferentialExpression-run_differential_expression}{\code{DifferentialExpression$run_differential_expression()}}
\item \href{#method-DifferentialExpression-ui}{\code{DifferentialExpression$ui()}}
\item \href{#method-DifferentialExpression-server}{\code{DifferentialExpression$server()}}
\item \href{#method-DifferentialExpression-clone}{\code{DifferentialExpression$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DifferentialExpression-new"></a>}}
\if{latex}{\out{\hypertarget{method-DifferentialExpression-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DifferentialExpression$new(dataHandling)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataHandling}}{An instance of DataHandling class.
Run differential expression analysis}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DifferentialExpression-run_differential_expression"></a>}}
\if{latex}{\out{\hypertarget{method-DifferentialExpression-run_differential_expression}{}}}
\subsection{Method \code{run_differential_expression()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DifferentialExpression$run_differential_expression(dset, ref_group, comp_group)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dset}}{Output from transform_xlsx()}

\item{\code{ref_group}}{String name of the reference group (e.g., "CDK4")}

\item{\code{comp_group}}{String name of the comparison group (e.g., "Bcell.only")}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List containing a SummarizedExperiment and the comparison config
Generate UI components
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DifferentialExpression-ui"></a>}}
\if{latex}{\out{\hypertarget{method-DifferentialExpression-ui}{}}}
\subsection{Method \code{ui()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DifferentialExpression$ui()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
shiny UI elements for tab panel
Server logic for differential expression module
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DifferentialExpression-server"></a>}}
\if{latex}{\out{\hypertarget{method-DifferentialExpression-server}{}}}
\subsection{Method \code{server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DifferentialExpression$server(input, output)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{Shiny input object}

\item{\code{output}}{Shiny output object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DifferentialExpression-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DifferentialExpression-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DifferentialExpression$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
