% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantProcessor.R
\name{VariantProcessor}
\alias{VariantProcessor}
\title{VariantProcessor Class}
\value{
An R6 object of class \code{VariantProcessor}.
}
\description{
This class handles the annotation of mutated regions from VCF files.
It processes VCF input files, annotates variants, generates peptide tables,
and manages the Shiny UI and server logic for interaction.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ens106}}{AnnotationHub object for Ensembl 106 genome annotations.}

\item{\code{asm}}{BSgenome object for genome assembly reference (e.g., hg38).}

\item{\code{rv}}{Reactive values container for internal state.}

\item{\code{test_mode_1}}{Logical flag to toggle test mode (default FALSE).}

\item{\code{rv_sheet}}{ReactiveValues object holding sheet and table data.}

\item{\code{dataHandling}}{Object handling data input/output operations.
Initialize VariantProcessor}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VariantProcessor-new}{\code{VariantProcessor$new()}}
\item \href{#method-VariantProcessor-process_vcf}{\code{VariantProcessor$process_vcf()}}
\item \href{#method-VariantProcessor-display_table}{\code{VariantProcessor$display_table()}}
\item \href{#method-VariantProcessor-ui}{\code{VariantProcessor$ui()}}
\item \href{#method-VariantProcessor-server}{\code{VariantProcessor$server()}}
\item \href{#method-VariantProcessor-clone}{\code{VariantProcessor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariantProcessor-new"></a>}}
\if{latex}{\out{\hypertarget{method-VariantProcessor-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantProcessor$new(dataHandling, test_mode_1 = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataHandling}}{External data handler object.}

\item{\code{test_mode_1}}{Logical, whether to use test data. Default FALSE.
Process VCF file to generate peptide annotation report}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariantProcessor-process_vcf"></a>}}
\if{latex}{\out{\hypertarget{method-VariantProcessor-process_vcf}{}}}
\subsection{Method \code{process_vcf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantProcessor$process_vcf(vcf_file = NULL, sample_selected = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vcf_file}}{Character path to VCF file. Default NULL for test data.}

\item{\code{sample_selected}}{Character sample name from VCF to process. Default NULL.
Display the annotated peptide tables with optional barcode input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariantProcessor-display_table"></a>}}
\if{latex}{\out{\hypertarget{method-VariantProcessor-display_table}{}}}
\subsection{Method \code{display_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantProcessor$display_table(output, input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{output}}{Shiny output object.}

\item{\code{input}}{Shiny input object.
UI method placeholder}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariantProcessor-ui"></a>}}
\if{latex}{\out{\hypertarget{method-VariantProcessor-ui}{}}}
\subsection{Method \code{ui()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantProcessor$ui()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to UI function (if any).
Server-side logic for variant calling workflow}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariantProcessor-server"></a>}}
\if{latex}{\out{\hypertarget{method-VariantProcessor-server}{}}}
\subsection{Method \code{server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantProcessor$server(input, output)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{Shiny input object.}

\item{\code{output}}{Shiny output object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariantProcessor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VariantProcessor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantProcessor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
